feat(vendors): isolate vendors backend controllers
